[ Cambio ]
Creacion de un widget que va a ser el encargado de mostrar los filed/inputs y selectores de personalizacion
